+++ projects/movies/src/app/pages/movie-list-page/movie-list-page.adapter.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { insert, RxState, selectSlice } from '@rx-angular/state';
-import { distinctUntilKeyChanged, EMPTY, filter, map, Observable, Subject, switchMap, withLatestFrom } from 'rxjs';
+import { distinctUntilKeyChanged, EMPTY, filter, map, Observable, switchMap, withLatestFrom } from 'rxjs';
 import { TMDBMovieModel } from '../../data-access/api/model/movie.model';
 import { getMovieCategory } from '../../data-access/api/resources/movie.resource';
 import { getDiscoverMovies } from '../../data-access/api/resources/discover.resource';
@@ -11,6 +11,7 @@ import { RouterState } from '../../shared/state/router.state';
 import { infiniteScrolled, InfiniteScrolleState, PaginationOptions } from '../../shared/utils/infinite-scroll/infinite-scrolled';
 import { RouterParams } from '../../shared/state/router-state.interface';
 import { PaginatedResult } from '../../shared/state/typings';
+import { getActions } from '../../shared/rxa-custom/actions';
 
 type MovieListRouterParams = Pick<RouterParams, 'type' | 'identifier'>
 type MovieListPageModel = InfiniteScrolleState<TMDBMovieModel> & MovieListRouterParams;
@@ -33,7 +34,7 @@ function getFetchByType(type: RouterParams['type']): (i: string, options?: TMDBP
 })
 export class MovieListPageAdapter extends RxState<MovieListPageModel> {
 
-  private readonly paginate$ = new Subject<void>();
+  private readonly actions = getActions<{ paginate: void }>();
 
   private readonly initialCategoryMovieList$ = (identifier: string) => this.movieState.select(
     selectSlice(['categoryMovies', 'categoryMoviesLoading']),
@@ -57,6 +58,8 @@ export class MovieListPageAdapter extends RxState<MovieListPageModel> {
     )
   );
 
+  readonly paginate = this.actions.paginate;
+
   constructor(
     private movieState: MovieState,
     private discoverState: DiscoverState,
@@ -64,7 +67,7 @@ export class MovieListPageAdapter extends RxState<MovieListPageModel> {
   ) {
     super();
 
-    const routerParamsFromPaginationTrigger$: Observable<MovieListRouterParams> = this.paginate$.pipe(
+    const routerParamsFromPaginationTrigger$: Observable<MovieListRouterParams> = this.actions.paginate$.pipe(
       withLatestFrom(this.routerState.routerParams$),
       map(([_, routerParams]) => routerParams)
     );
@@ -84,7 +87,7 @@ export class MovieListPageAdapter extends RxState<MovieListPageModel> {
             routerParamsFromPaginationTrigger$,
             // initial value
             (type === 'category' ? this.initialCategoryMovieList$(identifier) : this.initialDiscoverMovieList$(identifier)).pipe(
-             map((r) => ({ ...r, type, identifier }))
+              map((r) => ({ ...r, type, identifier }))
             )
           );
         })
@@ -99,7 +102,4 @@ export class MovieListPageAdapter extends RxState<MovieListPageModel> {
     );
   }
 
-  paginate() {
-    this.paginate$.next();
-  }
 }
